Doing an "open" followed by a "save as" when the file (not folder) is passed in as the starting point and no file is selected after doing a save results in a bad path.
eg:

When you open a transformation (VFS), it saves the path.
Ex: file:///c:/tmp/a.ktr

When you Save As another transformation, it that path to default the directory selection in the VfsBrowser.

But it sets the path of the file, not the directory!
Visually, the directory (file:///c:/tmp) appears to be selected, but internally the VfsBrowser thinks the .ktr is selected.

When you hit save, it tries to save to file:///c:/tmp/a.ktr/b.ktr. 

SEMINOLE-264